+++ RELEASE-NOTES.md
@@ -1,11 +1,14 @@
 ## RELEASE NOTES
+### Version 6.10.7-case-flags-v2-welsh-language-support-additional-rework
+**EUI-8116** Additional rework of Welsh language support
+
 ### Version 6.10.7-case-flags-v2-view-case-flags-ui-amendments
 **EUI-8069** Fix Case Flags table display to show selected language for "Language Interpreter" flag types (ported from Case Flags v1)
 **EUI-8070** Display "decision reason" (flag update comment) for "Not approved" flags to internal HMCTS users only
 **EUI-8071** Hide "Active flags" banner from external users
 
 ### Version 6.10.7-case-flags-legal-rep-comments-page-v3
-**EUI-7373** Reasonable Adjustments legal rep display and edit ‘Tell us why the support is no longer needed' screen
+**EUI-7373** Reasonable Adjustments legal rep display and edit 'Tell us why the support is no longer needed' screen
 
 ### Version 6.10.7-case-flags-welsh-language-support
 **EUI-7858** Ensure support for Welsh language toggle by storing flag description and comments entries in correct `_cy` fields, and displaying flag comments according to language selection